# Optimize startup times by avoiding loading fonts twice.