This change decreases editor startup time by ~0.44s (10%).

I believe this fixes a regression introduced by #61772 (so technically a `4.0` regression).

It should be possible to cherry-pick this change to earlier versions, but it's technically a behavior change (fonts that were previously loaded may error now, if an invalid font face index was provided). I'll let folks be the judge if that's worth it.

I tested loading fonts with an invalid font face index. It returns early from the function and prints an error, as expected.

## Explanation
When prompted to load a font, `TextServerAdv` first finds the number of available faces in the font by passing `-1` into `FT_Open_Face`. This feature is [documented](https://freetype.org/freetype2/docs/reference/ft2-face_creation.html#ft_open_face) in the function's description. Unfortunately, the function description does not mention that the function still fully loads the font. In effect, this means we've been loading fonts twice as often as needed. 

## Benchmark
I profiled a ~0.44s decrease in editor startup time:
```
 ❯ hyperfine -m25 -iw1 "bin/godot.macos.editor.arm64.master --path /Users/lukas/dev/godot/empty-game --editor --quit" "bin/godot.macos.editor.arm64 --path /Users/lukas/dev/godot/empty-game --editor --quit"
Benchmark 1: bin/godot.macos.editor.arm64.master --path /Users/lukas/dev/godot/empty-game --editor --quit
  Time (mean ± σ):      4.738 s ±  0.146 s    [User: 3.250 s, System: 0.400 s]
  Range (min … max):    4.357 s …  4.925 s    25 runs

Benchmark 2: bin/godot.macos.editor.arm64 --path /Users/lukas/dev/godot/empty-game --editor --quit
  Time (mean ± σ):      4.297 s ±  0.093 s    [User: 2.827 s, System: 0.394 s]
  Range (min … max):    4.133 s …  4.386 s    25 runs
```

## Addendum
Font loading is still a large contributor to editor startup time (6%). It may be possible to further decrease the amount of times fonts are loaded. But this would require a more invasive change.